+++ src/interfaces/ICommandBus.ts
@@ -1,15 +1,13 @@
 import type { ICommand } from './ICommand.ts';
-import type { IEventSet } from './IEventSet.ts';
-import type { IMessageMeta } from './IMessageMeta.ts';
 import type { IObservable } from './IObservable.ts';
 import type { IObserver } from './IObserver.ts';
 
 export interface ICommandBus extends IObservable {
-	send(commandType: string, aggregateId?: string, options?: { payload?: object, context?: object }):
-		Promise<IEventSet>;
+	send(command: ICommand): Promise<any>;
+	send(commandType: string, aggregateId?: string, options?: { payload?: object, context?: object }): Promise<any>;
 
-	sendRaw(command: ICommand, meta?: IMessageMeta):
-		Promise<IEventSet>;
+	/** @deprecated Use {@link send} */
+	sendRaw(command: ICommand): Promise<any>;
 }
 
 export interface ICommandHandler extends IObserver {