+++ src/rabbitmq/index.ts
@@ -1,2 +1,5 @@
+import './IContainer.ts';
+
 export * from './RabbitMqEventBus.ts';
 export * from './RabbitMqGateway.ts';
+export * from './RabbitMqCommandBus.ts';