+++ tests/unit/SagaEventHandler.test.ts
@@ -3,7 +3,6 @@ import {
 	SagaEventHandler,
 	InMemoryEventStorage,
 	EventStore,
-	CommandBus,
 	AbstractSaga,
 	InMemoryMessageBus,
 	EventDispatcher
@@ -33,7 +32,7 @@ const triggeringEvent = {
 
 describe('SagaEventHandler', function () {
 
-	let commandBus: CommandBus;
+	let commandBus: InMemoryMessageBus;
 	let eventStore: EventStore;
 	let sagaEventHandler: SagaEventHandler;
 	let eventStorage: InMemoryEventStorage;
@@ -42,7 +41,7 @@ describe('SagaEventHandler', function () {
 		const eventBus = new InMemoryMessageBus();
 		const eventDispatcher = new EventDispatcher({ eventBus });
 		eventStorage = new InMemoryEventStorage();
-		commandBus = new CommandBus({});
+		commandBus = new InMemoryMessageBus({});
 		eventStore = new EventStore({
 			eventStorageReader: eventStorage,
 			identifierProvider: eventStorage,